Apply pr fixes
