--- Sources/XcbeautifyLib/String+TeamCityEscaping.swift
@@ -1,15 +1,0 @@
-import Foundation
-
-extension String {
-    func teamCityEscaped() -> String {
-        // According to the documentation: https://www.jetbrains.com/help/teamcity/service-messages.html#Escaped+Values
-        return self
-            .replacingOccurrences(of: "|", with: "||")
-            .replacingOccurrences(of: "'", with: "|'")
-            .replacingOccurrences(of: "\n", with: "|n")
-            .replacingOccurrences(of: "\r", with: "|r")
-            .replacingOccurrences(of: "\\u{", with: "|u{") // Assuming the unicode format in Swift is \\u{NNNN}
-            .replacingOccurrences(of: "[", with: "|[")
-            .replacingOccurrences(of: "]", with: "|]")
-    }
-}
